[Fix] PostSummaryResponse LikeCount 필드 수정

+++ src/main/java/org/sopt/bofit/domain/post/dto/response/PostSummaryResponse.java
@@ -30,7 +30,7 @@ public record PostSummaryResponse (
         LocalDateTime createdAt,
 
         @Schema(description = "좋아요 수")
-        Long likeCount,
+        int likeCount,
 
         @Schema(description = "사용자의 좋아요 여부")
         boolean likedByCurrentUser
